+++ curvlinops/ekfac.py
@@ -6,13 +6,14 @@ from functools import partial
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 from einops import einsum, rearrange
-from torch import Generator, Tensor, cat
+from torch import Tensor, cat
 from torch.linalg import eigh
 from torch.nn import Conv2d, Module
 from torch.utils.hooks import RemovableHandle
 
 from curvlinops.kfac import FisherType, KFACLinearOperator
 from curvlinops.kfac_utils import extract_patches
+from curvlinops.utils import _seed_generator
 
 
 def compute_eigenvalue_correction_linear_weight_sharing(
@@ -22,7 +23,7 @@ def compute_eigenvalue_correction_linear_weight_sharing(
     aaT_eigvecs: Union[Tensor, None],
     _force_strategy: Optional[str] = None,
 ) -> Tensor:
-    r"""Computes eigenvalue corrections for a linear layer with weight sharing.
+    r"""Compute eigenvalue corrections for a linear layer with weight sharing.
 
     Chooses between two computational strategies depending on memory requirements.
 
@@ -246,14 +247,20 @@ class EKFACLinearOperator(KFACLinearOperator):
         _SUPPORTED_FISHER_TYPE: Tuple with supported Fisher types.
     """
 
-    _SUPPORTED_FISHER_TYPE: Tuple[FisherType] = (
+    _SUPPORTED_FISHER_TYPE: Tuple[FisherType, ...] = (
         FisherType.TYPE2,
         FisherType.MC,
         FisherType.EMPIRICAL,
     )
 
     @property
-    def representation(self):
+    def representation(self) -> Dict:
+        """Return EKFAC's internal representation (eigenvectors + corrected eigenvalues).
+
+        Returns:
+            A dictionary containing the eigenvectors of the input and gradient covariances
+            and the corrected eigenvalues for each module.
+        """
         if self._representation is None:
             input_covariances, gradient_covariances = self.compute_kronecker_factors()
             input_covariances_eigenvectors = self._eigenvectors_(input_covariances)
@@ -337,10 +344,8 @@ class EKFACLinearOperator(KFACLinearOperator):
             corrected_eigenvals = corrected_eigenvalues[mod_name]
 
             # bias and weights are treated jointly
-            if (
-                not self._separate_weight_and_bias
-                and "weight" in param_pos.keys()
-                and "bias" in param_pos.keys()
+            if not self._separate_weight_and_bias and {"weight", "bias"} == set(
+                param_pos.keys()
             ):
                 w_pos, b_pos = param_pos["weight"], param_pos["bias"]
                 # v denotes the free dimension for treating multiple vectors in parallel
@@ -422,9 +427,7 @@ class EKFACLinearOperator(KFACLinearOperator):
                 )
             )
 
-        if self._generator is None or self._generator.device != self.device:
-            self._generator = Generator(device=self.device)
-        self._generator.manual_seed(self._seed)
+        self._generator = _seed_generator(self._generator, self.device, self._seed)
 
         # loop over data set, computing the corrected eigenvalues
         for X, y in self._loop_over_data(desc="Eigenvalue correction"):
@@ -525,7 +528,7 @@ class EKFACLinearOperator(KFACLinearOperator):
             g = rearrange(g, "batch c o1 o2 -> batch o1 o2 c")
         g = rearrange(g, "batch ... d_out -> batch (...) d_out")
 
-        # We only need layer inputs to extracting information w.r.t. the weights
+        # We only need layer inputs to extract information w.r.t. the weights
         param_pos = self._mapping[module_name]
         a_required = "weight" in param_pos
 
@@ -549,11 +552,10 @@ class EKFACLinearOperator(KFACLinearOperator):
 
         # Compute correction for the loss scaling depending on the loss reduction used
         num_loss_terms = batch_size * self._num_per_example_loss_terms
-        # self._mc_samples will be 1 if fisher_type != FisherType.MC
         correction = {
-            "sum": 1.0 / self._mc_samples,
+            "sum": 1.0,
             "mean": num_loss_terms**2
-            / (self._N_data * self._mc_samples * self._num_per_example_loss_terms),
+            / (self._N_data * self._num_per_example_loss_terms),
         }[self._loss_func.reduction]
 
         # Compute the corrected eigenvalues for the EKFAC approximation
@@ -562,10 +564,9 @@ class EKFACLinearOperator(KFACLinearOperator):
         # ggT_eigenvectors always exists
         ggT_eigenvectors = gradient_covariances_eigenvectors[module_name]
 
-        if not self._separate_weight_and_bias and set(param_pos.keys()) == {
-            "weight",
-            "bias",
-        }:
+        if not self._separate_weight_and_bias and {"weight", "bias"} == set(
+            param_pos.keys()
+        ):
             a_augmented = cat([a, a.new_ones(*a.shape[:-1], 1)], dim=-1)
             eigencorrection = compute_eigenvalue_correction_linear_weight_sharing(
                 g, ggT_eigenvectors, a_augmented, aaT_eigenvectors
@@ -604,18 +605,18 @@ class EKFACLinearOperator(KFACLinearOperator):
         corrected_eigenvalues = self.representation["corrected_eigenvalues"]
 
         # Compute the trace using the corrected eigenvalues
-        trace = 0.0
+        _trace = 0.0
         for corrected_eigenvals in corrected_eigenvalues.values():
             if isinstance(corrected_eigenvals, dict):
                 for val in corrected_eigenvals.values():
-                    trace += val.sum()
+                    _trace += val.sum()
             else:
-                trace += corrected_eigenvals.sum()
+                _trace += corrected_eigenvals.sum()
 
-        return trace
+        return _trace
 
     def det(self) -> Tensor:
-        r"""Determinant of the EKFAC approximation.
+        r"""Compute the determinant of the EKFAC approximation.
 
         Will call ``compute_kronecker_factors`` and ``compute_eigenvalue_correction`` if
         either of them has not been called before.
@@ -626,20 +627,20 @@ class EKFACLinearOperator(KFACLinearOperator):
         corrected_eigenvalues = self.representation["corrected_eigenvalues"]
 
         # Compute the determinant using the corrected eigenvalues
-        det = 1.0
+        _det = 1.0
         for corrected_eigenvals in corrected_eigenvalues.values():
             if isinstance(corrected_eigenvals, dict):
                 for val in corrected_eigenvals.values():
-                    det *= val.prod()
+                    _det *= val.prod()
             else:
-                det *= corrected_eigenvals.prod()
+                _det *= corrected_eigenvals.prod()
 
-        return det
+        return _det
 
     def logdet(self) -> Tensor:
         r"""Log determinant of the EKFAC approximation.
 
-        More numerically stable than the ``det`` property.
+        More numerically stable than the ``det`` method.
         Will call ``compute_kronecker_factors`` and ``compute_eigenvalue_correction`` if
         either of them has not been called before.
 
@@ -649,15 +650,15 @@ class EKFACLinearOperator(KFACLinearOperator):
         corrected_eigenvalues = self.representation["corrected_eigenvalues"]
 
         # Compute the log determinant using the corrected eigenvalues
-        logdet = 0.0
+        _logdet = 0.0
         for corrected_eigenvals in corrected_eigenvalues.values():
             if isinstance(corrected_eigenvals, dict):
                 for val in corrected_eigenvals.values():
-                    logdet += val.log().sum()
+                    _logdet += val.log().sum()
             else:
-                logdet += corrected_eigenvals.log().sum()
+                _logdet += corrected_eigenvals.log().sum()
 
-        return logdet
+        return _logdet
 
     def frobenius_norm(self) -> Tensor:
         r"""Frobenius norm of the EKFAC approximation.
@@ -671,12 +672,12 @@ class EKFACLinearOperator(KFACLinearOperator):
         corrected_eigenvalues = self.representation["corrected_eigenvalues"]
 
         # Compute the Frobenius norm using the corrected eigenvalues
-        frobenius_norm = 0.0
+        _frobenius_norm = 0.0
         for corrected_eigenvals in corrected_eigenvalues.values():
             if isinstance(corrected_eigenvals, dict):
                 for val in corrected_eigenvals.values():
-                    frobenius_norm += val.square().sum()
+                    _frobenius_norm += val.square().sum()
             else:
-                frobenius_norm += corrected_eigenvals.square().sum()
+                _frobenius_norm += corrected_eigenvals.square().sum()
 
-        return frobenius_norm.sqrt()
+        return _frobenius_norm.sqrt()