+++ docs/examples/basic_usage/example_submatrices.py
@@ -1,5 +1,4 @@
-r"""
-Sub-matrices of linear operators
+r"""Sub-matrices of linear operators
 ================================
 
 This tutorial explains how to create linear operators that correspond to a sub-matrix